Broken FatFs multi-write support.
This has a partially implemented multi-write call. This works only for
files less than about 32K or so, which makes me think the issue is
probably something to do with cluster boundaries and how the caches switch
around when you hit them. I haven't been able to figure this out, so I'm
going to try switching to the much simpler Petit FatFs and trade my
problems out :)