Moving away from the "one byte pending" USART approach.
This was making my head hurt. The new approach is slightly slower but should be easier to understand.

This also leaves the RX part of the USART enabled during write ops. As I recall, this was an issue ages ago when I was experimenting with DMA transactions, which may come back again later.